refactor: Resolve AccountStatContext circular dependency
No jira.
