+++ src/Contexts.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const AccountStatContext = createContext({
+  hasConventionalSystems: true,
+  hasEdgeDevices: false,
+  hasBootcImages: false,
+});
