+++ src/components/InventoryTabs/ConventionalSystems/ConventionalSystemsTab.js
@@ -34,8 +34,8 @@ import useTableActions from './useTableActions';
 import useGlobalFilter from '../../filters/useGlobalFilter';
 import useOnRefresh from '../../filters/useOnRefresh';
 import useFeatureFlag from '../../../Utilities/useFeatureFlag';
-import { AccountStatContext } from '../../../Routes';
 import useWorkspaceFeatureFlag from '../../../Utilities/hooks/useWorkspaceFeatureFlag';
+import { AccountStatContext } from '../../../Contexts';
 
 const BulkDeleteButton = ({ selectedSystems, ...props }) => {
   const requiredPermissions = selectedSystems.map(({ groups }) =>
