+++ src/routes/InventoryComponents/HybridInventory.js
@@ -6,7 +6,7 @@ import { Bullseye, Spinner } from '@patternfly/react-core';
 import { useSearchParams } from 'react-router-dom';
 import { getSearchParams } from '../../constants';
 import useFeatureFlag from '../../Utilities/useFeatureFlag';
-import { AccountStatContext } from '../../Routes';
+import { AccountStatContext } from '../../Contexts';
 
 const ConventionalSystemsTab = lazy(() =>
   import(
