+++ src/routes/InventoryComponents/InventoryPageHeader.js
@@ -17,8 +17,8 @@ import {
 import { DesktopIcon } from '@patternfly/react-icons';
 import useFeatureFlag from '../../Utilities/useFeatureFlag';
 import { pageContents } from '../InventoryPage';
-import { AccountStatContext } from '../../Routes';
 import FontAwesomeImageIcon from '../../components/FontAwesomeImageIcon';
+import { AccountStatContext } from '../../Contexts';
 
 const InventoryContentToggle = ({ changeMainContent, mainContent }) => (
   <Split hasGutter>
